add removeMessage only for userOwner && add display message in chat not just from userOwner

+++ src/helpers/UI/ContextMenu.tsx
@@ -4,7 +4,8 @@ import axios from 'axios';
 import { useDispatch } from 'react-redux';
 import { useTypedSelector } from '../../hooks/useTypedSelector';
 import { urlChats } from '../../url/url';
-import { modalErrorState } from '../../store/reducers/modalErrorReducer';
+import { modalErrorState, modalErrorText } from '../../store/reducers/modalErrorReducer';
+import { userOwner } from '../userOwner';
 
 
 interface ContextMenuProps {
@@ -18,19 +19,36 @@ const ContextMenu: FC<ContextMenuProps> = ({ setOpenContextMenu, clientX, client
 
   const dispatch = useDispatch();
 
-  const openModalError = () => {
-      dispatch(modalErrorState(true));
+  const openModalErrorNoMessage = () => {
+    dispatch(modalErrorState(true));
+    dispatch(modalErrorText('Choose a message to delete'));
   };
- 
-  const { chosenMessage: id } = useTypedSelector(state => state.chosenMessage);
+
+  const openModalErrorNotYourMessage = () => {
+    dispatch(modalErrorState(true));
+    dispatch(modalErrorText('Choose your message to delete'));
+  };
+
+
+
+  const { chosenMessageId: id } = useTypedSelector(state => state.chosenMessage);
+  const { chosenMessageUserFrom: userFrom } = useTypedSelector(state => state.chosenMessage);
 
 
   const removeMessage = (event: React.MouseEvent<HTMLDivElement>) => {
-    setOpenContextMenu(false);
-    event.stopPropagation();
-    axios.delete(`${urlChats}/${id}`).then(() => {
-      onRemove();
-    }).catch(openModalError);
+    if (userFrom == userOwner || userFrom == null) {
+      setOpenContextMenu(false);
+      event.stopPropagation();
+      axios.delete(`${urlChats}/${id}`).then(() => {
+        onRemove();
+      }).catch(openModalErrorNoMessage);
+
+    }
+    else {
+      setOpenContextMenu(false);
+      event.stopPropagation();
+      openModalErrorNotYourMessage();
+    }
   };
